requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
langgraph==0.6.6
pydantic==2.11.7
python-dotenv==1.1.1
openai==1.30.1
